+++ packages/core-app-api/src/app/AppManager.tsx
@@ -82,8 +82,8 @@ import { resolveRouteBindings } from './resolveRouteBindings';
 import { BackstageRouteObject } from '../routing/types';
 
 type CompatiblePlugin =
-  | BackstagePlugin<any, any>
-  | (Omit<BackstagePlugin<any, any>, 'getFeatureFlags'> & {
+  | BackstagePlugin
+  | (Omit<BackstagePlugin, 'getFeatureFlags'> & {
       output(): Array<{ type: 'feature-flag'; name: string }>;
     });
 
@@ -145,12 +145,16 @@ function useConfigLoader(
 class AppContextImpl implements AppContext {
   constructor(private readonly app: AppManager) {}
 
-  getPlugins(): BackstagePlugin<any, any>[] {
+  getPlugins(): BackstagePlugin[] {
     return this.app.getPlugins();
   }
 
   getSystemIcon(key: string): IconComponent | undefined {
     return this.app.getSystemIcon(key);
+  }
+
+  getSystemIcons(): Record<string, IconComponent> {
+    return this.app.getSystemIcons();
   }
 
   getComponents(): AppComponents {
@@ -186,12 +190,16 @@ export class AppManager implements BackstageApp {
     this.apiFactoryRegistry = new ApiFactoryRegistry();
   }
 
-  getPlugins(): BackstagePlugin<any, any>[] {
-    return Array.from(this.plugins) as BackstagePlugin<any, any>[];
+  getPlugins(): BackstagePlugin[] {
+    return Array.from(this.plugins) as BackstagePlugin[];
   }
 
   getSystemIcon(key: string): IconComponent | undefined {
     return this.icons[key];
+  }
+
+  getSystemIcons(): Record<string, IconComponent> {
+    return this.icons;
   }
 
   getComponents(): AppComponents {
@@ -241,7 +249,7 @@ export class AppManager implements BackstageApp {
         validateRouteParameters(routing.paths, routing.parents);
         validateRouteBindings(
           routeBindings,
-          this.plugins as Iterable<BackstagePlugin<any, any>>,
+          this.plugins as Iterable<BackstagePlugin>,
         );
       }