+++ plugins/jenkins-backend/src/service/router.ts
@@ -28,12 +28,14 @@ import {
 import { getBearerTokenFromAuthorizationHeader } from '@backstage/plugin-auth-node';
 import { stringifyEntityRef } from '@backstage/catalog-model';
 
+/** @public */
 export interface RouterOptions {
   logger: Logger;
   jenkinsInfoProvider: JenkinsInfoProvider;
   permissions?: PermissionEvaluator | PermissionAuthorizer;
 }
 
+/** @public */
 export async function createRouter(
   options: RouterOptions,
 ): Promise<express.Router> {
@@ -64,12 +66,12 @@ export async function createRouter(
         request.header('authorization'),
       );
       const branch = request.query.branch;
-      let branchStr: string | undefined;
+      let branches: string[] | undefined;
 
       if (branch === undefined) {
-        branchStr = undefined;
+        branches = undefined;
       } else if (typeof branch === 'string') {
-        branchStr = branch;
+        branches = branch.split(/,/g);
       } else {
         // this was passed in as something weird -> 400
         // https://evanhahn.com/gotchas-with-express-query-parsing-and-how-to-avoid-them/
@@ -88,7 +90,7 @@ export async function createRouter(
         },
         backstageToken: token,
       });
-      const projects = await jenkinsApi.getProjects(jenkinsInfo, branchStr);
+      const projects = await jenkinsApi.getProjects(jenkinsInfo, branches);
 
       response.json({
         projects: projects,