+++ plugins/pagerduty/src/plugin.ts
@@ -19,15 +19,16 @@ import {
   createPlugin,
   createRouteRef,
   discoveryApiRef,
+  fetchApiRef,
   configApiRef,
   createComponentExtension,
-  identityApiRef,
 } from '@backstage/core-plugin-api';
 
 export const rootRouteRef = createRouteRef({
   id: 'pagerduty',
 });
 
+/** @public */
 export const pagerDutyPlugin = createPlugin({
   id: 'pagerduty',
   apis: [
@@ -36,14 +37,15 @@ export const pagerDutyPlugin = createPlugin({
       deps: {
         discoveryApi: discoveryApiRef,
         configApi: configApiRef,
-        identityApi: identityApiRef,
+        fetchApi: fetchApiRef,
       },
-      factory: ({ configApi, discoveryApi, identityApi }) =>
-        PagerDutyClient.fromConfig(configApi, discoveryApi, identityApi),
+      factory: ({ configApi, discoveryApi, fetchApi }) =>
+        PagerDutyClient.fromConfig(configApi, { discoveryApi, fetchApi }),
     }),
   ],
 });
 
+/** @public */
 export const EntityPagerDutyCard = pagerDutyPlugin.provide(
   createComponentExtension({
     name: 'EntityPagerDutyCard',