+++ plugins/stack-overflow-backend/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @backstage/plugin-stack-overflow-backend
 
+## 0.1.5-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-search-common@1.0.1-next.0
+
+## 0.1.4
+
+### Patch Changes
+
+- ea5631a8b2: Added API key as separate configuration
+
+## 0.1.4-next.0
+
+### Patch Changes
+
+- ea5631a8b2: Added API key as separate configuration
+
+## 0.1.3
+
+### Patch Changes
+
+- 52b4f796e3: app-config is now picked up properly.
+- Updated dependencies
+  - @backstage/plugin-search-common@1.0.0
+
+## 0.1.3-next.0
+
+### Patch Changes
+
+- 52b4f796e3: app-config is now picked up properly.
+- Updated dependencies
+  - @backstage/plugin-search-common@0.3.6-next.0
+
 ## 0.1.2
 
 ### Patch Changes